+++ Modules/Intent.Modules.VisualStudio.Projects/Templates/SqlServerDatabase/SqlProject/SqlProjectTemplatePartial.cs
@@ -4,12 +4,20 @@ using System.Linq;
 using System.Xml.Linq;
 using System.Xml.XPath;
 using Intent.Engine;
+using Intent.Eventing;
+using Intent.Exceptions;
 using Intent.Metadata.Models;
 using Intent.Modules.Common;
+using Intent.Modules.Common.CSharp.Nuget;
 using Intent.Modules.Common.Templates;
 using Intent.Modules.Common.VisualStudio;
+using Intent.Modules.Constants;
 using Intent.Modules.VisualStudio.Projects.Api;
+using Intent.Modules.VisualStudio.Projects.Events;
+using Intent.Modules.VisualStudio.Projects.FactoryExtensions;
 using Intent.Modules.VisualStudio.Projects.Sync;
+using Intent.Modules.VisualStudio.Projects.Templates;
+using Intent.Modules.VisualStudio.Projects.Templates.SqlServerDatabase.SqlProject.Content;
 using Intent.RoslynWeaver.Attributes;
 using Intent.Templates;
 
@@ -19,14 +27,101 @@ using Intent.Templates;
 namespace Intent.Modules.VisualStudio.Projects.Templates.SqlServerDatabase.SqlProject
 {
     [IntentManaged(Mode.Ignore)]
-    public partial class SqlProjectTemplate : VisualStudioProjectTemplateBase<SQLServerDatabaseProjectModel>
+    public partial class SqlProjectTemplate : IntentFileTemplateBase<SQLServerDatabaseProjectModel>, IVisualStudioProjectTemplate
     {
-        [IntentManaged(Mode.Fully)]
-        public const string TemplateId = "Intent.VisualStudio.Projects.SqlServerDatabase.SqlProject";
+        [IntentManaged(Mode.Fully)] public const string TemplateId = "Intent.VisualStudio.Projects.SqlServerDatabase.SqlProject";
+
+        private string _fileContent;
 
         [IntentManaged(Mode.Merge, Signature = Mode.Fully)]
         public SqlProjectTemplate(IOutputTarget outputTarget, SQLServerDatabaseProjectModel model) : base(TemplateId, outputTarget, model)
         {
+        }
+
+        public override void OnCreated()
+        {
+            base.OnCreated();
+            ExecutionContext.EventDispatcher.Publish(new VisualStudioProjectCreatedEvent(this));
+        }
+
+        public string ProjectId => Model.Id;
+        public string Name => Model.Name;
+        public string FilePath => FileMetadata.GetFilePath();
+        IVisualStudioProject IVisualStudioProjectTemplate.Project => Model;
+
+        public string LoadContent()
+        {
+            var change = ExecutionContext.ChangeManager.FindChange(FilePath);
+            if (change != null)
+            {
+                return change.Content;
+            }
+
+            if (_fileContent == null)
+            {
+                TryGetExistingFileContent(out _fileContent);
+            }
+
+            return _fileContent;
+        }
+
+        public void UpdateContent(string content, ISoftwareFactoryEventDispatcher sfEventDispatcher)
+        {
+            var targetContent = XDocument.Parse(content).ToString().ReplaceLineEndings();
+            var existingContent = LoadContent().ReplaceLineEndings();
+
+            if (existingContent == targetContent)
+            {
+                return;
+            }
+
+            var change = ExecutionContext.ChangeManager.FindChange(FilePath);
+            if (change != null)
+            {
+                change.ChangeContent(content, content);
+                return;
+            }
+
+            sfEventDispatcher.Publish(new SoftwareFactoryEvent(SoftwareFactoryEvents.OverwriteFileCommand, new Dictionary<string, string>
+            {
+                ["FullFileName"] = FilePath,
+                ["Context"] = ToString(),
+                ["Content"] = content
+            }));
+        }
+
+        public IEnumerable<NuGetInstall> RequestedNugetPackageInstalls() => [];
+        public IEnumerable<string> GetTargetFrameworks() => Model.TargetFrameworkVersion();
+
+        public override string TransformText()
+        {
+            if (IsSdkStyle())
+            {
+                return SqlProjectSdkContent.Generate(
+                    Model.Name,
+                    Model.Id,
+                    Model.GetSQLServerDatabaseProject().Version());
+            }
+
+            return SqlProjectFrameworkContent.Generate(Model.Name, Model.Id);
+        }
+
+        private bool IsSdkStyle()
+        {
+            var hasSqlServerDatabaseProject = Model.HasSQLServerDatabaseProject();
+            var hasNETFrameworkSettings = Model.HasNETFrameworkSettings();
+
+            if (!hasSqlServerDatabaseProject && !hasNETFrameworkSettings)
+            {
+                throw new ElementException(
+                    Model.InternalElement,
+                    $"SQL Server Database Project '{Model.Name}' must have either a 'SQL Server Database Project' or '.NET Framework Settings' stereotype applied.");
+            }
+
+            if (!hasSqlServerDatabaseProject)
+                return false;
+
+            return Model.GetSQLServerDatabaseProject().ProjectType().IsSDK();
         }
 
         public override ITemplateFileConfig GetTemplateFileConfig()
@@ -37,19 +132,39 @@ namespace Intent.Modules.VisualStudio.Projects.Templates.SqlServerDatabase.SqlPr
             );
         }
 
-        protected override string ApplyAdditionalTransforms(string existingFileOrTransformTextContent)
+        public override string RunTemplate()
         {
-            var xml = new ProjectFileXml(existingFileOrTransformTextContent);
+            var hadExistingContent = TryGetExistingFileContent(out var existingFileContent);
 
-            ApplySqlCmdVariables(xml);
+            var content = hadExistingContent && IsMatchingStructure(existingFileContent)
+                ? existingFileContent
+                : TransformText();
 
-            return xml.Document.ToStringUTF8();
+            if (!IsSdkStyle())
+            {
+                var xml = new ProjectFileXml(content);
+                ApplySqlCmdVariables(xml);
+                content = xml.Document.ToStringUTF8();
+            }
+
+            var doc = XDocument.Parse(content);
+
+            return !hadExistingContent || !XmlHelper.IsSemanticallyTheSame(existingFileContent, doc)
+                ? doc.ToFormattedProjectString()
+                : existingFileContent;
+        }
+
+        private bool IsMatchingStructure(string content)
+        {
+            var doc = XDocument.Parse(content);
+            var isSdkContent = doc.Root?.Name.Namespace == XNamespace.None;
+            return IsSdkStyle() == isSdkContent;
         }
 
         /// <summary>
         /// SQL projects should never have NuGet packages.
         /// </summary>
-        public override IEnumerable<INugetPackageInfo> RequestedNugetPackages() => [];
+        public IEnumerable<INugetPackageInfo> RequestedNugetPackages() => [];
 
         private void ApplySqlCmdVariables(ProjectFileXml xml)
         {